🏷️ customize global object type
