+++ packages/core/src/domain/connectivity/connectivity.ts
@@ -1,17 +1,5 @@
+import type { EffectiveType, NetworkInterface } from '../../browser/browser.types'
 import { globalObject } from '../../tools/globalObject'
-
-export type NetworkInterface = 'bluetooth' | 'cellular' | 'ethernet' | 'none' | 'wifi' | 'wimax' | 'other' | 'unknown'
-export type EffectiveType = 'slow-2g' | '2g' | '3g' | '4g'
-
-interface BrowserNavigator extends Navigator {
-  connection?: NetworkInformation
-}
-
-export interface NetworkInformation {
-  type?: NetworkInterface
-  effectiveType?: EffectiveType
-  saveData: boolean
-}
 
 export interface Connectivity {
   status: 'connected' | 'not_connected'
@@ -21,7 +9,7 @@ export interface Connectivity {
 }
 
 export function getConnectivity(): Connectivity {
-  const navigator = globalObject.navigator as BrowserNavigator
+  const navigator = globalObject.navigator
 
   return {
     status: navigator.onLine ? 'connected' : 'not_connected',
